pyproject.toml
[tool.poetry]
name = "fastapitest"
version = "0.1.0"
description = ""
authors = ["CHAMOMILE <CHAMOM2LES@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
jinja2 = "^3.1.3"
python-multipart = "^0.0.9"
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
black = "^24.4.2"
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
